Rename Celbridge.CommonViews to Celbridge.Views
