+++ Celbridge/Celbridge.Views/Pages/MainPage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="Celbridge.CommonViews.Pages.MainPage"
+    x:Class="Celbridge.Views.Pages.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Celbridge.CommonViews.Pages"
+    xmlns:local="using:Celbridge.Views.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
