+++ Celbridge/Celbridge.Views/Pages/NewProjectPage.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="Celbridge.CommonViews.Pages.SettingsPage"
+    x:Class="Celbridge.Views.Pages.NewProjectPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Celbridge.CommonViews.Pages"
+    xmlns:local="using:Celbridge.Views.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -12,7 +12,7 @@
         <StackPanel HorizontalAlignment="Center"
                     VerticalAlignment="Center"
                     Spacing="6">
-            <TextBlock Text="Settings Page"/>
+            <TextBlock Text="New Project Page"/>
         </StackPanel>
     </Grid>
 </Page>
