+++ Celbridge/Celbridge.Views/Pages/StartPage.xaml
@@ -1,18 +1,20 @@
 ﻿<Page
-    x:Class="Celbridge.CommonViews.Pages.NewProjectPage"
+    x:Class="Celbridge.Views.Pages.StartPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Celbridge.CommonViews.Pages"
+    xmlns:local="using:Celbridge.Views.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
     DataContext="{x:Bind ViewModel}">
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <TextBlock x:Uid="/Resources/TestString" />
         <StackPanel HorizontalAlignment="Center"
                     VerticalAlignment="Center"
                     Spacing="6">
-            <TextBlock Text="New Project Page"/>
+            <Button Content="{x:Bind OpenWorkspace}" 
+                    Command="{x:Bind ViewModel.OpenWorkspacePageCommand}" />
         </StackPanel>
     </Grid>
 </Page>
