+++ Celbridge/Celbridge.Views/Resources/AppResources.xaml
@@ -9,7 +9,7 @@
         <ToolkitResources xmlns="using:Uno.Toolkit.UI" />
 
         <!-- Add resource dictionaries here -->
-        <ResourceDictionary Source="ms-appx:///Celbridge.CommonViews/Resources/Colors.xaml" />
+        <ResourceDictionary Source="ms-appx:///Celbridge.Views/Resources/Colors.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
 </ResourceDictionary>
