+++ Celbridge/Celbridge.Views/Resources/Colors.xaml
