+++ Celbridge/Celbridge.Views/UserControls/TitleBar.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="Celbridge.CommonViews.UserControls.TitleBar"
+    x:Class="Celbridge.Views.UserControls.TitleBar"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
