+++ Celbridge/CoreExtensions/Celbridge.Workspace/Views/WorkspacePage.xaml
@@ -2,7 +2,7 @@
     x:Class="Celbridge.Workspace.Views.WorkspacePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Celbridge.CommonViews.Pages"
+    xmlns:local="using:Celbridge.Views.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
